requirements/django.txt
Django==3.2.23
psycopg2 >= 2.7.3.1
protobuf==3.20.2
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-django==0.42b0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-distro==0.42b0
opentelemetry-instrumentation-requests==0.42b0
opentelemetry-instrumentation-psycopg2==0.42b0
python-decouple==3.8
psycopg2-binary==2.8.5
django-postgres-extra==2.0.3
requests==2.31.0
gunicorn==20.0.4
